Step 10: the one source on the GPU — the recipe crate compiled to SPIR-V runs the shipped function on Metal, byte for byte
Ruling F8 decision 5 (rust-gpu, one crate, two compilations), step (c)(2)-A.

THE SHELL (crates/recipe-gpu, vd-recipe-gpu): one compute entry point, relief_columns, that reads
its bindings and calls vd_recipe::height::relief_of_table — no arithmetic of its own; built by
cargo-gpu (rust-gpu 0.10.0-alpha.1 at rev 7fa56ad6, its pinned nightly 2026-05-22, the Int64
capability) through `just recipe-gpu`; excluded from the stable workspace, which declares the
spirv target arch for the cfg check.

THE RULES THE GPU COMPILER SET, now in the recipe: the octave sum is an index loop (no slice
iterator) and gains a fixed-size-table form (no runtime-length slice of a local array); the
gradient and basis tables are 32-bit words (no 8-bit integer without a capability); the 128-bit
cell-count reciprocal is gated off the GPU target (the charter carries the word); the terrain
crate cross-pins its octave cap to the recipe's; wgpu ingests SPIR-V on the render crate.

MEASURED (bench part 4, crates/bins/examples/integer_bench.rs, slice_08_integer_bench.md): the
module loaded by naga into Metal on the Apple M4 Pro — 0 of 3 936 256 columns differ from the
CPU's call of the same function; the CPU 710 ms on one core, the GPU 285 ms and 348 ms on two
passes with the upload and the readback (the transcription's 56–127 ms names the cost lever:
the per-thread copy of the octave table). DX12 and Vulkan stay unmeasured until another machine
runs the same module.

+++ crates/recipe/src/bend.rs
@@ -44,15 +44,17 @@ pub const K3: Gi = Gi::new(71_030_470_521);
 const _: () = assert!(K1.raw() + K2.raw() + K3.raw() == 1 << DIR_BITS);
 
 /// A face's basis: the normal, the `u` axis (along `i`) and the `v` axis (along `j`), each an axis
-/// with one component of ±1 — the address format's table, as `vd_seed::bend::BASIS` has it.
+/// with one component of ±1 — the address format's table, as `vd_seed::bend::BASIS` has it. The
+/// components are 32-bit words: the GPU target carries no 8-bit integer without a capability of
+/// its own.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct Basis {
     /// The face's outward normal.
-    pub n: [i8; 3],
+    pub n: [i32; 3],
     /// The axis a cell's `i` counts along.
-    pub u: [i8; 3],
+    pub u: [i32; 3],
     /// The axis a cell's `j` counts along.
-    pub v: [i8; 3],
+    pub v: [i32; 3],
 }
 
 /// THE FACE BASIS TABLE, by face index: +X, −X, +Y, −Y, +Z, −Z.
@@ -97,12 +99,14 @@ pub const BASIS: [Basis; 6] = [
 /// `i64::MIN` and `2⁶⁴/1` truncates to zero, and either word would bend a face inside out. `2⁶⁴/3`
 /// still fits the word, so three is the floor. The clamp is a GUARD, not a path: a face of a body on
 /// the ladder never has fewer than one chunk's width of cells (62), because the ladder's own snap
-/// keeps every rung at least that wide.
+/// keeps every rung at least that wide. The GPU target does not compile this function (a 128-bit
+/// division) and never runs it: the charter carries the word.
 #[must_use]
 #[allow(
     clippy::integer_division,
     reason = "computed once per body on the CPU, never in a kernel"
 )]
+#[cfg(not(target_arch = "spirv"))]
 pub const fn inv_n_of(n_l: u32) -> Gi {
     let n = if n_l < INV_MIN_CELLS {
         INV_MIN_CELLS
